Moved IHaveIBeenPwnedRestClientto the interfaces folder
